+++ docs/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+### 2023-02-11 ⚙️
+
+* Rewrite [daixin](profiles/daixin.md) parser to add description
+  
 ### 2023-02-11 💰
 
 * Add crypto wallet when available in group profile